split parser
