+++ parser/float.go
@@ -1,0 +1,23 @@
+package parser
+
+import (
+	"fmt"
+	"strconv"
+
+	"github.com/flipez/rocket-lang/ast"
+)
+
+func (p *Parser) parseFloat() ast.Expression {
+	lit := &ast.Float{Token: p.curToken}
+
+	value, err := strconv.ParseFloat(p.curToken.Literal, 64)
+	if err != nil {
+		msg := fmt.Sprintf("could not parse %q as float", p.curToken.Literal)
+		p.errors = append(p.errors, msg)
+		return nil
+	}
+
+	lit.Value = value
+
+	return lit
+}
