+++ parser/hash.go
@@ -1,0 +1,36 @@
+package parser
+
+import (
+	"github.com/flipez/rocket-lang/ast"
+	"github.com/flipez/rocket-lang/token"
+)
+
+func (p *Parser) parseHash() ast.Expression {
+	hash := &ast.Hash{Token: p.curToken}
+	hash.Pairs = make(map[ast.Expression]ast.Expression)
+
+	for !p.peekTokenIs(token.RBRACE) {
+		p.nextToken()
+		key := p.parseExpression(LOWEST)
+
+		if !p.expectPeek(token.COLON) {
+			return nil
+		}
+
+		p.nextToken()
+		value := p.parseExpression(LOWEST)
+
+		hash.Pairs[key] = value
+
+		if !p.peekTokenIs(token.RBRACE) && !p.expectPeek(token.COMMA) {
+			return nil
+		}
+
+	}
+
+	if !p.expectPeek(token.RBRACE) {
+		return nil
+	}
+
+	return hash
+}
