+++ parser/identifier.go
@@ -1,0 +1,9 @@
+package parser
+
+import (
+	"github.com/flipez/rocket-lang/ast"
+)
+
+func (p *Parser) parseIdentifier() ast.Expression {
+	return &ast.Identifier{Token: p.curToken, Value: p.curToken.Literal}
+}
