+++ parser/import.go
@@ -1,0 +1,24 @@
+package parser
+
+import (
+	"github.com/flipez/rocket-lang/ast"
+	"github.com/flipez/rocket-lang/token"
+)
+
+func (p *Parser) parseImport() ast.Expression {
+	expression := &ast.Import{Token: p.curToken}
+
+	if !p.expectPeek(token.LPAREN) {
+		return nil
+	}
+
+	p.nextToken()
+
+	expression.Name = p.parseExpression(LOWEST)
+
+	if !p.expectPeek(token.RPAREN) {
+		return nil
+	}
+
+	return expression
+}
