+++ parser/integer.go
@@ -1,0 +1,23 @@
+package parser
+
+import (
+	"fmt"
+	"strconv"
+
+	"github.com/flipez/rocket-lang/ast"
+)
+
+func (p *Parser) parseInteger() ast.Expression {
+	lit := &ast.Integer{Token: p.curToken}
+
+	value, err := strconv.ParseInt(p.curToken.Literal, 0, 64)
+	if err != nil {
+		msg := fmt.Sprintf("could not parse %q as integer", p.curToken.Literal)
+		p.errors = append(p.errors, msg)
+		return nil
+	}
+
+	lit.Value = value
+
+	return lit
+}
