+++ parser/parser.go
@@ -2,8 +2,6 @@ package parser
 
 import (
 	"fmt"
-	"path/filepath"
-	"strconv"
 
 	"github.com/flipez/rocket-lang/ast"
 	"github.com/flipez/rocket-lang/lexer"
@@ -78,7 +76,7 @@ func New(l *lexer.Lexer, imports map[string]struct{}) *Parser {
 	p.registerPrefix(token.FUNCTION, p.parseFunction)
 	p.registerPrefix(token.STRING, p.parseString)
 	p.registerPrefix(token.LBRACKET, p.parseArray)
-	p.registerPrefix(token.LBRACE, p.parseHashLiteral)
+	p.registerPrefix(token.LBRACE, p.parseHash)
 	p.registerPrefix(token.IMPORT, p.parseImport)
 
 	p.infixParseFns = make(map[token.TokenType]infixParseFn)
@@ -92,7 +90,7 @@ func New(l *lexer.Lexer, imports map[string]struct{}) *Parser {
 	p.registerInfix(token.PERIOD, p.parseMethodCall)
 	p.registerInfix(token.LT, p.parseInfix)
 	p.registerInfix(token.GT, p.parseInfix)
-	p.registerInfix(token.LPAREN, p.parseCallExpression)
+	p.registerInfix(token.LPAREN, p.parseCall)
 	p.registerInfix(token.LBRACKET, p.parseIndex)
 
 	// read two tokens, so curToken and peekToken are both set
@@ -122,130 +120,6 @@ func (p *Parser) nextToken() {
 	p.peekToken = p.l.NextToken()
 }
 
-func (p *Parser) ParseProgram() (*ast.Program, map[string]struct{}) {
-	program := &ast.Program{}
-	program.Statements = []ast.Statement{}
-
-	for p.curToken.Type != token.EOF {
-		stmt := p.parseStatement()
-		if stmt != nil {
-			program.Statements = append(program.Statements, stmt)
-
-			if expStmt, ok := stmt.(*ast.ExpressionStatement); ok {
-				if importExpr, ok := expStmt.Expression.(*ast.Import); ok {
-					implicitVarName := filepath.Base(importExpr.Name.String())
-					p.imports[implicitVarName] = struct{}{}
-				}
-			}
-		}
-		p.nextToken()
-	}
-
-	return program, p.imports
-}
-
-func (p *Parser) parseHashLiteral() ast.Expression {
-	hash := &ast.Hash{Token: p.curToken}
-	hash.Pairs = make(map[ast.Expression]ast.Expression)
-
-	for !p.peekTokenIs(token.RBRACE) {
-		p.nextToken()
-		key := p.parseExpression(LOWEST)
-
-		if !p.expectPeek(token.COLON) {
-			return nil
-		}
-
-		p.nextToken()
-		value := p.parseExpression(LOWEST)
-
-		hash.Pairs[key] = value
-
-		if !p.peekTokenIs(token.RBRACE) && !p.expectPeek(token.COMMA) {
-			return nil
-		}
-
-	}
-
-	if !p.expectPeek(token.RBRACE) {
-		return nil
-	}
-
-	return hash
-}
-
-func (p *Parser) parseIndex(left ast.Expression) ast.Expression {
-	exp := &ast.Index{Token: p.curToken, Left: left}
-
-	p.nextToken()
-	exp.Index = p.parseExpression(LOWEST)
-
-	if !p.expectPeek(token.RBRACKET) {
-		return nil
-	}
-
-	return exp
-}
-
-func (p *Parser) parseArray() ast.Expression {
-	array := &ast.Array{Token: p.curToken}
-
-	array.Elements = p.parseExpressionList(token.RBRACKET)
-
-	return array
-}
-
-func (p *Parser) parseExpressionList(end token.TokenType) []ast.Expression {
-	list := []ast.Expression{}
-
-	if p.peekTokenIs(end) {
-		p.nextToken()
-		return list
-	}
-
-	p.nextToken()
-	list = append(list, p.parseExpression(LOWEST))
-
-	for p.peekTokenIs(token.COMMA) {
-		p.nextToken()
-		p.nextToken()
-		list = append(list, p.parseExpression(LOWEST))
-	}
-
-	if !p.expectPeek(end) {
-		return nil
-	}
-
-	return list
-}
-
-func (p *Parser) parseString() ast.Expression {
-	return &ast.String{Token: p.curToken, Value: p.curToken.Literal}
-}
-
-func (p *Parser) parseStatement() ast.Statement {
-	switch p.curToken.Type {
-	case token.RETURN:
-		return p.parseReturn()
-	default:
-		return p.parseExpressionStatement()
-	}
-}
-
-func (p *Parser) parseReturn() *ast.Return {
-	stmt := &ast.Return{Token: p.curToken}
-
-	p.nextToken()
-
-	stmt.ReturnValue = p.parseExpression(LOWEST)
-
-	if p.peekTokenIs(token.SEMICOLON) {
-		p.nextToken()
-	}
-
-	return stmt
-}
-
 func (p *Parser) parseExpressionStatement() *ast.ExpressionStatement {
 	stmt := &ast.ExpressionStatement{Token: p.curToken}
 
@@ -256,75 +130,6 @@ func (p *Parser) parseExpressionStatement() *ast.ExpressionStatement {
 	}
 
 	return stmt
-}
-
-func (p *Parser) parseExpression(precedence int) ast.Expression {
-	prefix := p.prefixParseFns[p.curToken.Type]
-	if prefix == nil {
-		p.noPrefixParseFnError(p.curToken.Type)
-		return nil
-	}
-	leftExp := prefix()
-
-	for !p.peekTokenIs(token.SEMICOLON) && precedence < p.peekPrecedence() {
-		infix := p.infixParseFns[p.peekToken.Type]
-		if infix == nil {
-			return leftExp
-		}
-
-		p.nextToken()
-
-		leftExp = infix(leftExp)
-	}
-
-	return leftExp
-}
-
-func (p *Parser) parseIdentifier() ast.Expression {
-	return &ast.Identifier{Token: p.curToken, Value: p.curToken.Literal}
-}
-
-func (p *Parser) parseInteger() ast.Expression {
-	lit := &ast.Integer{Token: p.curToken}
-
-	value, err := strconv.ParseInt(p.curToken.Literal, 0, 64)
-	if err != nil {
-		msg := fmt.Sprintf("could not parse %q as integer", p.curToken.Literal)
-		p.errors = append(p.errors, msg)
-		return nil
-	}
-
-	lit.Value = value
-
-	return lit
-}
-
-func (p *Parser) parseFloat() ast.Expression {
-	lit := &ast.Float{Token: p.curToken}
-
-	value, err := strconv.ParseFloat(p.curToken.Literal, 64)
-	if err != nil {
-		msg := fmt.Sprintf("could not parse %q as float", p.curToken.Literal)
-		p.errors = append(p.errors, msg)
-		return nil
-	}
-
-	lit.Value = value
-
-	return lit
-}
-
-func (p *Parser) parsePrefix() ast.Expression {
-	expression := &ast.Prefix{
-		Token:    p.curToken,
-		Operator: p.curToken.Literal,
-	}
-
-	p.nextToken()
-
-	expression.Right = p.parseExpression(PREFIX)
-
-	return expression
 }
 
 func (p *Parser) curTokenIs(t token.TokenType) bool {
@@ -366,256 +171,4 @@ func (p *Parser) curPrecedence() int {
 	}
 
 	return LOWEST
-}
-
-func (p *Parser) parseInfix(left ast.Expression) ast.Expression {
-	expression := &ast.Infix{
-		Token:    p.curToken,
-		Operator: p.curToken.Literal,
-		Left:     left,
-	}
-
-	precedence := p.curPrecedence()
-	p.nextToken()
-	expression.Right = p.parseExpression(precedence)
-
-	return expression
-}
-
-func (p *Parser) parseBoolean() ast.Expression {
-	return &ast.Boolean{Token: p.curToken, Value: p.curTokenIs(token.TRUE)}
-}
-
-func (p *Parser) parseGroupedExpression() ast.Expression {
-	p.nextToken()
-
-	exp := p.parseExpression(LOWEST)
-
-	if !p.expectPeek(token.RPAREN) {
-		return nil
-	}
-
-	return exp
-}
-
-func (p *Parser) parseIf() ast.Expression {
-	expression := &ast.If{Token: p.curToken}
-	if !p.expectPeek(token.LPAREN) {
-		return nil
-	}
-
-	p.nextToken()
-	expression.Condition = p.parseExpression(LOWEST)
-
-	if !p.expectPeek(token.RPAREN) {
-		return nil
-	}
-
-	if p.peekTokenIs(token.LBRACE) {
-		p.nextToken()
-	}
-	expression.Consequence = p.parseBlock()
-
-	if p.curTokenIs(token.RBRACE) {
-		p.nextToken()
-	}
-
-	if p.curTokenIs(token.ELSE) {
-
-		if p.peekTokenIs(token.LBRACE) {
-			p.nextToken()
-		}
-		expression.Alternative = p.parseBlock()
-	}
-	return expression
-}
-
-func (p *Parser) parseBlock() *ast.Block {
-	block := &ast.Block{Token: p.curToken}
-	block.Statements = []ast.Statement{}
-
-	p.nextToken()
-
-	for !p.curTokenIs(token.RBRACE) && !p.curTokenIs(token.EOF) && !p.curTokenIs(token.END) && !p.curTokenIs(token.ELSE) {
-		stmt := p.parseStatement()
-		if stmt != nil {
-			block.Statements = append(block.Statements, stmt)
-		}
-
-		p.nextToken()
-	}
-
-	return block
-}
-
-func (p *Parser) parseFunction() ast.Expression {
-	lit := &ast.Function{Token: p.curToken}
-
-	if !p.peekTokenIs(token.LPAREN) {
-		p.nextToken()
-
-		lit.Name = p.curToken.Literal
-	}
-
-	if !p.expectPeek(token.LPAREN) {
-		return nil
-	}
-
-	lit.Parameters = p.parseFunctionParameters()
-
-	if !p.expectPeek(token.LBRACE) {
-		return nil
-	}
-
-	lit.Body = p.parseBlock()
-
-	return lit
-}
-
-func (p *Parser) parseFunctionParameters() []*ast.Identifier {
-	identifiers := []*ast.Identifier{}
-
-	if p.peekTokenIs(token.RPAREN) {
-		p.nextToken()
-		return identifiers
-	}
-
-	p.nextToken()
-
-	ident := &ast.Identifier{Token: p.curToken, Value: p.curToken.Literal}
-	identifiers = append(identifiers, ident)
-
-	for p.peekTokenIs(token.COMMA) {
-		p.nextToken()
-		p.nextToken()
-		ident := &ast.Identifier{Token: p.curToken, Value: p.curToken.Literal}
-		identifiers = append(identifiers, ident)
-	}
-
-	if !p.expectPeek(token.RPAREN) {
-		return nil
-	}
-
-	return identifiers
-}
-
-func (p *Parser) parseCallExpression(callable ast.Expression) ast.Expression {
-	exp := &ast.Call{Token: p.curToken, Callable: callable}
-	exp.Arguments = p.parseExpressionList(token.RPAREN)
-	return exp
-}
-
-func (p *Parser) parseCallArguments() []ast.Expression {
-	args := []ast.Expression{}
-
-	if p.peekTokenIs(token.RPAREN) {
-		p.nextToken()
-		return args
-	}
-
-	p.nextToken()
-	args = append(args, p.parseExpression(LOWEST))
-
-	for p.peekTokenIs(token.COMMA) {
-		p.nextToken()
-		p.nextToken()
-		args = append(args, p.parseExpression(LOWEST))
-	}
-
-	if !p.expectPeek(token.RPAREN) {
-		return nil
-	}
-
-	return args
-}
-
-func (p *Parser) parseMethodCall(obj ast.Expression) ast.Expression {
-	if _, ok := p.imports[obj.String()]; ok {
-		p.expectPeek(token.IDENT)
-		index := &ast.String{Token: p.curToken, Value: p.curToken.Literal}
-		return &ast.Index{Left: obj, Index: index}
-	}
-
-	methodCall := &ast.ObjectCall{Token: p.curToken, Object: obj}
-	p.nextToken()
-	name := p.parseIdentifier()
-	p.nextToken()
-	methodCall.Call = p.parseCallExpression(name)
-	return methodCall
-}
-
-func (p *Parser) parseForEach() ast.Expression {
-	expression := &ast.Foreach{Token: p.curToken}
-
-	p.nextToken()
-	expression.Ident = p.curToken.Literal
-
-	if p.peekTokenIs(token.COMMA) {
-
-		p.nextToken()
-
-		if !p.peekTokenIs(token.IDENT) {
-			p.errors = append(p.errors, fmt.Sprintf("second argument to foreach must be ident, got %v", p.peekToken))
-			return nil
-		}
-		p.nextToken()
-
-		expression.Index = expression.Ident
-		expression.Ident = p.curToken.Literal
-
-	}
-
-	if !p.expectPeek(token.IN) {
-		return nil
-	}
-	p.nextToken()
-
-	expression.Value = p.parseExpression(LOWEST)
-	if expression.Value == nil {
-		return nil
-	}
-
-	p.nextToken()
-	expression.Body = p.parseBlock()
-
-	return expression
-}
-
-func (p *Parser) parseAssignExpression(name ast.Expression) ast.Expression {
-	stmt := &ast.Assign{Token: p.curToken}
-	if n, ok := name.(*ast.Identifier); ok {
-		stmt.Name = n
-	} else {
-		msg := fmt.Sprintf("expected assign token to be IDENT, got %s instead", name.TokenLiteral())
-		p.errors = append(p.errors, msg)
-	}
-	p.nextToken()
-	stmt.Value = p.parseExpression(LOWEST)
-	return stmt
-}
-
-func (p *Parser) parseImport() ast.Expression {
-	expression := &ast.Import{Token: p.curToken}
-
-	if !p.expectPeek(token.LPAREN) {
-		return nil
-	}
-
-	p.nextToken()
-
-	expression.Name = p.parseExpression(LOWEST)
-
-	if !p.expectPeek(token.RPAREN) {
-		return nil
-	}
-
-	return expression
-}
-
-func (p *Parser) parseDotNotationExpression(expression ast.Expression) ast.Expression {
-	p.expectPeek(token.IDENT)
-
-	index := &ast.String{Token: p.curToken, Value: p.curToken.Literal}
-
-	return &ast.Index{Left: expression, Index: index}
 }
