+++ parser/return.go
@@ -1,0 +1,20 @@
+package parser
+
+import (
+	"github.com/flipez/rocket-lang/ast"
+	"github.com/flipez/rocket-lang/token"
+)
+
+func (p *Parser) parseReturn() *ast.Return {
+	stmt := &ast.Return{Token: p.curToken}
+
+	p.nextToken()
+
+	stmt.ReturnValue = p.parseExpression(LOWEST)
+
+	if p.peekTokenIs(token.SEMICOLON) {
+		p.nextToken()
+	}
+
+	return stmt
+}
