+++ parser/statement.go
@@ -1,0 +1,15 @@
+package parser
+
+import (
+	"github.com/flipez/rocket-lang/ast"
+	"github.com/flipez/rocket-lang/token"
+)
+
+func (p *Parser) parseStatement() ast.Statement {
+	switch p.curToken.Type {
+	case token.RETURN:
+		return p.parseReturn()
+	default:
+		return p.parseExpressionStatement()
+	}
+}
